Add float-time render API for sub-frame keyframe sampling
Propagate float frameNo through the renderer and layer-update path so the
keyframe interpolator can be sampled between integer frame indices. Exposes:

  C++:  Animation::renderSyncAtFrameF(double frameNo, Surface, bool keepAspectRatio)
  C:    lottie_animation_render_at_frame_f(anim, double, buffer, w, h, stride)

Lets callers render smoothly on displays whose refresh rate exceeds the
animation's authored framerate (e.g. a 25 fps Lottie on a 120 Hz device).

The existing size_t-based render entry points are unchanged.

+++ src/lottie/lottiefiltermodel.h
@@ -366,7 +366,7 @@ public:
     CapStyle  capStyle() const { return this->model()->capStyle(); }
     JoinStyle joinStyle() const { return this->model()->joinStyle(); }
     bool      hasDashInfo() const { return this->model()->hasDashInfo(); }
-    void      getDashInfo(int frameNo, std::vector<float>& result) const
+    void      getDashInfo(float frameNo, std::vector<float>& result) const
     {
         return this->model()->getDashInfo(frameNo, result);
     }
@@ -431,7 +431,7 @@ class Filter<model::Trim>: public FilterBase<model::Trim>
 public:
     Filter(model::Trim* model) : FilterBase<model::Trim>(model) {}
 
-    model::Trim::Segment segment(int frameNo) const
+    model::Trim::Segment segment(float frameNo) const
     {
         if (this->hasFilter(rlottie::Property::TrimStart)) {
             this->model_->updateTrimStartValue(this->filter()->value(rlottie::Property::TrimStart, frameNo));
